+++ pkg/modules/autostart/autostart_test.go
@@ -10,7 +10,7 @@ import (
 	"howett.net/plist"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/gnzdotmx/ishinobu/pkg/utils"
 )
