+++ pkg/modules/coreanalytics/coreanalytics_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/stretchr/testify/assert"
 )
