+++ pkg/modules/firefox/firefox_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/gnzdotmx/ishinobu/pkg/utils"
 	"github.com/stretchr/testify/assert"
 )
