+++ pkg/modules/netcfgplists/netcfgplists_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/gnzdotmx/ishinobu/pkg/utils"
 	"github.com/stretchr/testify/require"
 	"howett.net/plist"
