+++ pkg/modules/quicklook/quicklook_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 
 	"github.com/stretchr/testify/assert"
 )
