+++ pkg/modules/safari/safari_test.go
@@ -12,7 +12,7 @@ import (
 	"howett.net/plist"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/gnzdotmx/ishinobu/pkg/utils"
 
 	_ "github.com/mattn/go-sqlite3"
