--- pkg/modules/spotlight/coverage.out
@@ -1,31 +1,0 @@
-mode: set
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:25.13,30.2 2 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:32.44,34.2 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:36.51,38.2 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:40.62,47.35 2 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:47.35,49.17 2 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:49.17,51.12 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:54.3,54.30 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:54.30,56.18 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:56.18,58.5 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:62.2,62.12 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:65.90,69.33 3 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:69.33,70.39 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:70.39,71.28 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:71.28,73.10 2 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:79.2,81.16 3 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:81.16,83.3 1 1
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:86.2,87.16 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:87.16,89.3 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:91.2,92.16 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:92.16,94.3 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:97.2,97.45 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:97.45,99.10 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:99.10,101.4 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:103.3,111.62 7 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:111.62,113.18 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:113.18,115.13 2 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:117.4,117.39 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:120.3,130.17 4 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:130.17,132.4 1 0
-github.com/gnzdotmx/ishinobu/pkg/modules/spotlight/spotlight.go:135.2,135.12 1 0
