+++ pkg/modules/systeminfo/systeminfo_test.go
@@ -11,7 +11,7 @@ import (
 	"howett.net/plist"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/gnzdotmx/ishinobu/pkg/utils"
 )
