+++ pkg/modules/users/users_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/gnzdotmx/ishinobu/pkg/mod"
-	"github.com/gnzdotmx/ishinobu/pkg/modules/testutils"
+	"github.com/gnzdotmx/ishinobu/pkg/testutils"
 	"github.com/gnzdotmx/ishinobu/pkg/utils"
 
 	"github.com/stretchr/testify/assert"
